Added additional custom date field
This field is for the datetime that the blog post is published. This is
different to the created and modified PW fields. At the moment I am
populating this with the post_date_gmt - is this appropriate?

As well as being a good idea in general, I also wanted this field
because Kongondo's blog module requires it is filled out. Although for
compatibility with that, the config setting needs to be changed to
blog_date.